# Improve SocketAddress.ToString() performance