Following up on request/feedback at https://github.com/dotnet/corefx/pull/30521#discussion_r287352510.

Microbenchmark:
```C#
using System;
using System.Net;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromTypes(new[] { typeof(Program) }).Run(args);

    private readonly SocketAddress _addr = new IPEndPoint(IPAddress.Parse("40.112.72.205"), 443).Serialize();

    [Benchmark]
    public string SAToString() => _addr.ToString(); 
}
```

|     Method |     Mean |    Error |   StdDev |   Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
| Before     | 528.7 ns | 10.67 ns | 24.73 ns | 517.1 ns | 0.1183 |     - |     - |     496 B |
| After        | 346.6 ns | 4.733 ns | 3.695 ns | 0.0362 |     - |     - |     - |     152 B |